Một số thay đổi nhỏ
Đưa mute_spin vào extract cho gọn
Loại bỏ delete_index, pydub_convert
Thêm 1 số điều kiện kiểm tra tệp tin huấn luyện
Thêm pretrainG = pretrainD = None vào training
Chỉnh sửa split log, đổi sang sample rate choice
Thêm optimizer
Loại bỏ flatten khỏi audio effects, thêm volume vào pydub load
chỉnh sửa resample
chỉnh sửa sang cách trích xuất dữ liệu mới
sửa lại extract để tăng tốc độ
chỉnh sửa fairseq chỉ giữ lại mô hình
sửa lại dtype cũ của audioldm2
thêm 1 except để tắt cudnn trong rmvpe
thêm stonemask vào swipe
đổi tên WORLD_WRAPPER -> WORLD

+++ main/library/audioldm2/models.py
@@ -257,7 +257,7 @@ class AudioLDM2(Pipeline):
         if not torch.is_tensor(timesteps):
             is_mps = sample.device.type == "mps"
 
-            dtype = (torch.float16 if is_mps else torch.float32) if isinstance(timestep, float) else (torch.int16 if is_mps else torch.int32)
+            dtype = (torch.float32 if is_mps else torch.float64) if isinstance(timestep, float) else (torch.int32 if is_mps else torch.int64)
 
             timesteps = torch.tensor([timesteps], dtype=dtype, device=sample.device)
         elif len(timesteps.shape) == 0: timesteps = timesteps[None].to(sample.device)
